introduce rule
